Move nodes into nodes.h and write wacky list() template function
